# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yoga API documentation",
        "contact": {
            "url": "http://t.me/kosynka"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://127.0.0.1:8000/",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/api/v1/user/": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Create user",
                "description": "This can only be done by the logged in user.",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Created user object",
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Список ввсех пользователей"
                    }
                }
            }
        },
...
    },
...
}